@@ -3,7 +3,7 @@
 @{
     ViewData["Title"] = "About";
 }
-<h2>@ViewData["Title"]</h2>
-<h3>@Model.Message</h3>
+<h1 class="card-title">@ViewData["Title"]</h1>
+<h2 class="card-title">@Model.Message</h2>
 
 <p>Use this area to provide additional information.</p>
