@@ -3,8 +3,8 @@
 @{
     ViewData["Title"] = "Contact";
 }
-<h2>@ViewData["Title"]</h2>
-<h3>@Model.Message</h3>
+<h1 class="card-title">@ViewData["Title"]</h1>
+<h2 class="card-title">@Model.Message</h2>
 
 <address>
     One Microsoft Way<br />
