@@ -3,21 +3,20 @@
 @{
     ViewData["Title"] = "Custom Application Parts";
 }
-<h2>@ViewData["Title"]</h2>
+<h2 class="card-title">@ViewData["Title"]</h2>
 <p>List all WebParts are loaded current Application Domain</p>
 
-<form method="post">
-    <table class="table">
-        <thead>
-            <tr>
-                <th>Application Part Assembly</th>
-                <th>Type Name</th>
-                <th>API Name</th>
-            </tr>
-        </thead>
-        <tbody>
-            @foreach (var appPart in Model.LoadedParts)
-            {
+<table class="table">
+    <thead>
+        <tr>
+            <th>Application Part Assembly</th>
+            <th>Type Name</th>
+            <th>API Name</th>
+        </tr>
+    </thead>
+    <tbody>
+        @foreach (var appPart in Model.LoadedParts)
+        {
             <tr>
                 <td>@appPart.AssemblyName</td>
                 <td>@appPart.TypeName</td>
@@ -26,7 +25,6 @@
                     <a href="../api/@appPart.ApiName" target="_blank">Test</a>
                 </td>
             </tr>
-            }
-        </tbody>
-    </table>
-</form>
+        }
+    </tbody>
+</table>
