@@ -3,6 +3,6 @@
 @{
     ViewData["Title"] = "Privacy Policy";
 }
-<h2>@ViewData["Title"]</h2>
+<h1 class="card-title">@ViewData["Title"]</h1>
 
 <p>Use this page to detail your site's privacy policy.</p>
