@@ -15,10 +15,14 @@ namespace ITSolution.Framework.Core.Server.BaseClasses.Repository
             _optionsBuilder = new DbContextOptionsBuilder();
             if (EnvironmentInformation.DatabaseType == DatabaseType.MSSQL)
                 _optionsBuilder.UseSqlServer(EnvironmentManager.Configuration.ConnectionString);
-            else
+            else if(EnvironmentInformation.DatabaseType == DatabaseType.Oracle)
             {
                 ITSOracleConfiguration.ConfigureDataSources();
-                _optionsBuilder.UseOracle(EnvironmentManager.Configuration.ConnectionString);
+                //_optionsBuilder.UseOracle(EnvironmentManager.Configuration.ConnectionString);
+            }
+            else if(EnvironmentInformation.DatabaseType == DatabaseType.SQLITE)
+            {
+                _optionsBuilder.UseSqlite(EnvironmentInformation.ConnectionString);
             }
         }
 
