@@ -1,5 +1,5 @@
 ﻿using ITSolution.Framework.BaseClasses;
-using Oracle.ManagedDataAccess.Client;
+//using Oracle.ManagedDataAccess.Client;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -13,7 +13,7 @@ namespace ITSolution.Framework.Core.Server.BaseClasses.Repository
         /// </summary>
         public static void ConfigureDataSources()
         {
-            OracleConfiguration.OracleDataSources.Add(EnvironmentInformation.DefaultConnectionName, EnvironmentInformation.TnsAtpSource);
+            //OracleConfiguration.OracleDataSources.Add(EnvironmentInformation.DefaultConnectionName, EnvironmentInformation.TnsAtpSource);
             //OracleConfiguration.WalletLocation = EnvironmentInformation.WalletLocation;
         }
     }
